portfolio-chatbot/backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
beautifulsoup4
fastembed
chromadb
groq
numpy
pydantic
PyMuPDF
python-docx
